Refactor LMIApp class into its own file
